feat: gut pi-embedded agent tools (web, browser, image, TTS) (#66)
Remove 11 pi-embedded agent tool source files, 15 test files, and 2
cascading test files from src/agents/tools/ (~8,500 lines). These tools
provided web fetch/search, browser control, image analysis, TTS, and
nested agent execution for OpenClaw's in-process LLM engine.
RemoteClaw's CLI agents have native equivalents via MCP or built-in
tooling — these are dead code.

- Delete 11 source files: web-fetch, web-fetch-utils,
  web-fetch-visibility, web-search, web-shared, web-tools, browser-tool,
  browser-tool.schema, image-tool, image-tool.helpers, tts-tool
- Delete 15 associated test files and 2 cascading test files
  (pi-tools browser test, config web-search-provider test)
- Inline coerceImageAssistantText into media-understanding/providers/image.ts
- Remove deleted tool imports and creation calls from openclaw-tools.ts
- Clean up stale vi.mock() in test helpers (fast-tool-stubs, fast-core-tools)
- Update pi-tools action-enum test to remove browser from expected tools

Keep agent-step.ts: issue proposed full deletion, but runAgentStep has a
live consumer in sessions-send-tool.a2a.ts.
